Allow labelOptions in DynamicLabels to affect each other
Since a common use-case for this is units, the newly added
"options_metadata" option allows you to mark the options as belonging to
a particular type.  Then a change in one will affect the others going
forward.

For example, you can use it to change between imperial and metric units.