Add a custom condition delegate
When determining whether to show a question, add a new operation called
"custom".  If "custom" is the operation, the calculaton of whether to
show the question is outsourced to a delegate.  Answers relevant to the
"ids" are passed into the delegate, as well as a dictionary of [String:
Any] called "extra" that can be specified in the json.